+++ packages/tswebextension/src/lib/common/utils/logger.ts
@@ -1,22 +1,70 @@
+import { LogLevelEnum, LogLevelType, LogLevelName } from '../configuration';
+import { appContext } from '../../mv2/background/context';
+// TODO (v.zhelvis) move app context to common and make it generic.
+
+const DEFAULT_VERBOSE_FLAG = true;
+const DEFAULT_LOG_LEVEL: LogLevelType = LogLevelEnum.enum.Error;
+
+/**
+ * Number presentation of log levels. Order is important. Higher number, more messages to be visible.
+ */
+const enum LogLevelWeight {
+    Error = 1,
+    Warn,
+    Info,
+    Debug,
+}
+
 /**
  * Logger interface.
  */
-interface LoggerInterface {
+export interface LoggerInterface {
     error(message?: string): void;
     warn(message?: string): void;
     debug(message?: string): void;
     info(message?: string): void;
 }
 
-export interface FlexibleLoggerInterface extends LoggerInterface {
-    setVerbose(value: boolean): void;
-}
-
 /**
  * Export logger implementation.
  */
-export class Logger implements FlexibleLoggerInterface {
-    private verbose = false;
+export class Logger implements LoggerInterface {
+    /**
+     * Gets app verbose status.
+     *
+     * TODO (v.zhelvis) remove eslint rule after passing appContext the right way.
+     *
+     * @returns App verbose status.
+     */
+    // eslint-disable-next-line class-methods-use-this
+    private get verbose(): boolean {
+        return appContext.configuration?.verbose ?? DEFAULT_VERBOSE_FLAG;
+    }
+
+    /**
+     * Gets app log level.
+     *
+     * TODO (v.zhelvis) remove eslint rule after passing appContext the right way.
+     *
+     * @returns Log level.
+     */
+    // eslint-disable-next-line class-methods-use-this
+    private get logLevel(): LogLevelWeight {
+        const logLevelString = appContext.configuration?.logLevel ?? DEFAULT_LOG_LEVEL;
+
+        switch (logLevelString) {
+            case LogLevelEnum.enum.Error:
+                return LogLevelWeight.Error;
+            case LogLevelEnum.enum.Warn:
+                return LogLevelWeight.Warn;
+            case LogLevelEnum.enum.Info:
+                return LogLevelWeight.Info;
+            case LogLevelEnum.enum.Debug:
+                return LogLevelWeight.Debug;
+            default:
+                throw new Error(`Logger only supports following levels: ${[Object.values(LogLevelName).join(', ')]}`);
+        }
+    }
 
     private loggerImpl: LoggerInterface;
 
@@ -28,20 +76,10 @@ export class Logger implements FlexibleLoggerInterface {
     constructor(loggerImpl: LoggerInterface = console) {
         this.loggerImpl = loggerImpl;
 
-        this.setVerbose = this.setVerbose.bind(this);
         this.error = this.error.bind(this);
         this.warn = this.warn.bind(this);
         this.debug = this.debug.bind(this);
         this.info = this.info.bind(this);
-    }
-
-    /**
-     * Sets verbose mode.
-     *
-     * @param value Boolean flag.
-     */
-    public setVerbose(value: boolean): void {
-        this.verbose = value;
     }
 
     /**
@@ -50,7 +88,9 @@ export class Logger implements FlexibleLoggerInterface {
      * @param message Log message.
      */
     public error(message?: string): void {
-        this.loggerImpl.error(message);
+        if (this.logLevel >= LogLevelWeight.Error) {
+            this.loggerImpl.error(message);
+        }
     }
 
     /**
@@ -59,7 +99,7 @@ export class Logger implements FlexibleLoggerInterface {
      * @param message Log message.
      */
     public warn(message?: string): void {
-        if (this.verbose) {
+        if (this.verbose && this.logLevel >= LogLevelWeight.Warn) {
             this.loggerImpl.warn(message);
         }
     }
@@ -70,7 +110,7 @@ export class Logger implements FlexibleLoggerInterface {
      * @param message Log message.
      */
     public debug(message?: string): void {
-        if (this.verbose) {
+        if (this.verbose && this.logLevel >= LogLevelWeight.Debug) {
             this.loggerImpl.debug(message);
         }
     }
@@ -81,7 +121,7 @@ export class Logger implements FlexibleLoggerInterface {
      * @param message Log message.
      */
     public info(message?: string): void {
-        if (this.verbose) {
+        if (this.verbose && this.logLevel >= LogLevelWeight.Info) {
             this.loggerImpl.info(message);
         }
     }