+++ packages/tswebextension/src/lib/mv2/background/context.ts
@@ -4,6 +4,8 @@ import type { ConfigurationMV2Context } from './configuration';
  * Top level app context storage.
  *
  * This context is needed to share data between other modules without cyclic dependencies.
+ *
+ * TODO (v.zhelvis) move app context to common and make it generic.
  */
 export class AppContext {
     /**