+++ packages/tswebextension/CHANGELOG.md
@@ -7,6 +7,15 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 <!-- TODO: manually add compare links for version to the end of the file -->
 <!-- e.g. [0.1.2]: https://github.com/AdguardTeam/tsurlfilter/compare/tswebextension-v0.1.1...tswebextension-v0.1.2 -->
 
+## [0.2.2] - 2023-05-29
+
+### Added
+- new `logLevel`optional property for MV2 configuration to control logging levels.
+
+### Changed
+- `verbose` MV2 configuration property is now optional.
+
+
 ## [0.2.0] - 2023-05-23
 
 ### Added